Skip waking layout reactions when nothing moved
- reconcilePositions returns the previous record when no item moved and no
  key was added or removed, so an identical relayout reuses the same object
  and hits reanimated's === dedup instead of waking every per-item derived.
- Guard the grid overriddenCellDimensions write so an unchanged column width
  no longer wakes every cell's animated style on container resize.
- Move reconcilePositions to utils/layout.ts; keep equality.ts to predicates.
- activeItemPosition now only seeds the initial mutable; live active updates
  flow through the CommonValuesProvider dispatcher.
- Return the runOnUI runner directly from the registration effect cleanup.
